updated simulator.py to read directories from config.ini
added parameter SIMULATOR_IMG_DIR in config.ini and updated simulator.py to read INPUT_DIR and OUTPUT_DIR from this file. In this way the only file to be modified to run COLMAP_SLAM in a different environment is the config.ini (which is now excluded from git tracking to not interfere when working in multiple people on the same repo.